Suggested changes

+++ include/thga.h
@@ -21,13 +21,13 @@ s32 THGA_GetRemaining(TwoHeadGfxArena* thga);
 Gfx* THGA_GetHead(TwoHeadGfxArena* thga);
 void THGA_SetHead(TwoHeadGfxArena* thga, Gfx* newHead);
 void* THGA_GetTail(TwoHeadGfxArena* thga);
-Gfx* THGA_AllocDisplayList(TwoHeadGfxArena* thga, u32 num);
+Gfx* THGA_AllocDisplayList(TwoHeadGfxArena* thga, size_t num);
 Gfx* THGA_AllocGfx(TwoHeadGfxArena* thga);
 Gfx* THGA_AllocGfx2(TwoHeadGfxArena* thga);
 void* THGA_AllocTail(TwoHeadGfxArena* thga, size_t size);
-Mtx* THGA_AllocMtxArray(TwoHeadGfxArena* thga, u32 num);
+Mtx* THGA_AllocMtxArray(TwoHeadGfxArena* thga, size_t num);
 Mtx* THGA_AllocMtx(TwoHeadGfxArena* thga);
-Vtx* THGA_AllocVtxArray(TwoHeadGfxArena* thga, u32 num);
+Vtx* THGA_AllocVtxArray(TwoHeadGfxArena* thga, size_t num);
 Vtx* THGA_AllocVtx(TwoHeadGfxArena* thga);
 
 #endif
