+++ src/code/TwoHeadGfxArena.c
@@ -5,6 +5,8 @@
  *
  * Display list commands are allocated from the head while other graphics data such as matrices and vertices are
  * allocated from the tail end.
+ * 
+ * @see TwoHeadArena.c
  */
 #include "global.h"
 
@@ -41,25 +43,30 @@ void* THGA_GetTail(TwoHeadGfxArena* thga) {
 }
 
 /**
- * Allocates a display list of `num` Gfx commands to the head of the Two Head GFX Arena.
+ * Allocates a display list of `num` Gfx commands to the head of the Two Head Gfx Arena.
  */
-Gfx* THGA_AllocDisplayList(TwoHeadGfxArena* thga, u32 num) {
+Gfx* THGA_AllocDisplayList(TwoHeadGfxArena* thga, size_t num) {
     return THA_AllocHead(&thga->tha, num * sizeof(Gfx));
 }
 
 /**
- * Allocates a single Gfx command to the head of the Two Head GFX Arena.
+ * Allocates a single Gfx command to the head of the Two Head Gfx Arena.
  */
 Gfx* THGA_AllocGfx(TwoHeadGfxArena* thga) {
     return THGA_AllocDisplayList(thga, 1);
 }
 
+/**
+ * Identical to `THGA_AllocGfx`
+ * 
+ * @see THGA_AllocGfx
+ */
 Gfx* THGA_AllocGfx2(TwoHeadGfxArena* thga) {
     return THGA_AllocGfx(thga);
 }
 
 /**
- * Allocates to the end of the Two Head GFX Arena. Intended for data complementary to the display lists such as
+ * Allocates to the end of the Two Head Gfx Arena. Intended for data complementary to the display lists such as
  * matrices and vertices that are only needed for a single graphics task.
  */
 void* THGA_AllocTail(TwoHeadGfxArena* thga, size_t size) {
@@ -67,28 +74,28 @@ void* THGA_AllocTail(TwoHeadGfxArena* thga, size_t size) {
 }
 
 /**
- * Allocates `num` matrices to the tail end of the Two Head GFX Arena.
+ * Allocates `num` matrices to the tail end of the Two Head Gfx Arena.
  */
-Mtx* THGA_AllocMtxArray(TwoHeadGfxArena* thga, u32 num) {
+Mtx* THGA_AllocMtxArray(TwoHeadGfxArena* thga, size_t num) {
     return THGA_AllocTail(thga, num * sizeof(Mtx));
 }
 
 /**
- * Allocates a matrix to the tail end of the Two Head GFX Arena.
+ * Allocates a matrix to the tail end of the Two Head Gfx Arena.
  */
 Mtx* THGA_AllocMtx(TwoHeadGfxArena* thga) {
     return THGA_AllocTail(thga, sizeof(Mtx));
 }
 
 /**
- * Allocates `num` vertices to the tail end of the Two Head GFX Arena.
+ * Allocates `num` vertices to the tail end of the Two Head Gfx Arena.
  */
-Vtx* THGA_AllocVtxArray(TwoHeadGfxArena* thga, u32 num) {
+Vtx* THGA_AllocVtxArray(TwoHeadGfxArena* thga, size_t num) {
     return THGA_AllocTail(thga, num * sizeof(Vtx));
 }
 
 /**
- * Allocates a vertex to the tail end of the Two Head GFX Arena.
+ * Allocates a vertex to the tail end of the Two Head Gfx Arena.
  */
 Vtx* THGA_AllocVtx(TwoHeadGfxArena* thga) {
     return THGA_AllocTail(thga, sizeof(Vtx));
